update structure and move logic into modals
