+++ packages/modals/components/AppModals/index.test.jsx
@@ -1,7 +1,16 @@
 import React from 'react';
 import ShallowRenderer from 'react-test-renderer/shallow';
+import { Provider } from 'react-redux';
 
+import AppModalsConnected from '../../index';
 import AppModals from './index';
+
+const fakeStore = state => ({
+  default: () => {},
+  subscribe: () => {},
+  dispatch: () => {},
+  getState: () => ({ ...state }),
+});
 
 const renderer = new ShallowRenderer();
 
@@ -17,6 +26,15 @@ describe('Modals', () => {
     const visibleModals = { showUpgradeModal: true };
     const rendered = renderer.render(
       <AppModals {...visibleModals} />
+    );
+    expect(rendered).toMatchSnapshot();
+  });
+  it('renders a connected upgrade modal', () => {
+    const visibleModals = { showUpgradeModal: true };
+    const rendered = renderer.render(
+      <Provider store={fakeStore({ modals: visibleModals })}>
+        <AppModalsConnected />
+      </Provider>
     );
     expect(rendered).toMatchSnapshot();
   });
