+++ packages/modals/reducer.js
@@ -1,0 +1,36 @@
+import keyWrapper from '@bufferapp/keywrapper';
+
+export const initialState = {
+  showUpgradeModal: false,
+};
+
+export const actionTypes = keyWrapper('MODALS', {
+  SHOW_UPGRADE_MODAL: 0,
+  HIDE_UPGRADE_MODAL: 0,
+});
+
+export default (state = initialState, action) => {
+  switch (action.type) {
+    case actionTypes.SHOW_UPGRADE_MODAL:
+      return {
+        ...state,
+        showUpgradeModal: true,
+      };
+    case actionTypes.HIDE_UPGRADE_MODAL:
+      return {
+        ...state,
+        showUpgradeModal: false,
+      };
+    default:
+      return state;
+  }
+};
+
+export const actions = {
+  showUpgradeModal: () => ({
+    type: actionTypes.SHOW_UPGRADE_MODAL,
+  }),
+  hideUpgradeModal: () => ({
+    type: actionTypes.HIDE_UPGRADE_MODAL,
+  }),
+};
